Adds ReceiptTrackerFactory and deletes the two instances of ReceiptTrackerService
